requirements/web.txt
# Requirements for the web frontend (eg. not on Lambda)

-r base.txt

django-allauth==0.36.0
django-bouncy==0.2.6
django-loginas==0.3.4
django-reflinks==0.5.0
django-webpack-loader==0.6.0

boto3==1.9.9
mailchimp3==3.0.5
psycopg2==2.7.4
uwsgi==2.0.17
unidecode==1.0.22

numpy
scikit-learn[alldeps]

# stripe / paypal
dj-stripe==1.2.2
dj-paypal==0.9.0
